Created About Page
Used FlexBox Layout to put image of me and text side by side.
Made hyperlinks on nav bar direct user to the headers by the same name in the website. Also added ability to download pdfs from the documents directory with the click of a button.

Problems to be solved: when page is shrinked, navbar elements spill, and icon appears out of place. Moreover, image is squished when the screen size is diminished.