Contact Form UI + fixed navbar icon
Nearly completed the Contact form's UI; however, noticed that the input boxes do not appear to conform to 20px padding of the contactFormBox class div.
Also, adjusted icon on navbar so that its distance from the leftmost border of the screen changes according to the window size.
Lastly, temporarily (or perhaps indefinitely) removed the 'Recommended Books' section of the website.